+++ changelog.md
@@ -1,2 +1,4 @@
 ### v1.0.2 (Sept. ?, 2019)
-* Added Freewind Post, Thousand Needles NPCs
+* Added Thousand Needles NPCs
+* Added Camp Mojache, Feralas NPCs
+* Added Tanaris NPCs